refactor: move some duplicated sorted dict methods in the base trait
